Introduce the callback 'emit'
That callback is intended to be used for custom
writing. If defined, the process will call it
instead of fwrite. If not defined, the process
calls fwrite.

To implement custom write function, a client can
call 'fmustach' with a 'file' being a custom pointer
with cast. Then the client implement 'emit' and 'put'
its way, casting the 'file' back to its custom type.
This can be NULL for example.

CAUTION: when using custom write, it is important
to also define the callback 'partial'. If you don't
do that, the process of mustach will call your
custom function with a real FILE!

Fixes issue #11

Signed-off-by: José Bollo <jobol@nonadev.net>