Introduce the callback 'get'
The original idea was to provide a simple implmentation
and then to avoid stuff like dealing with sbuf.

It was cool but had serious drawback:
 - implementation of partial was hidden
 - the output was to standard FILE only (with hidden fwrite)
 - its confused 2 distincts behaviours:
   * retrieving values
   * emiting data

This change fixes that last issues by providing
a clean separation between retrieving and emiting.

It makes the implementation of callback 'put' optionnal
by providing a default implementation that calls 'get'
followed by 'emit'.

Signed-off-by: José Bollo <jobol@nonadev.net>